+++ packages/context/src/repository/index.ts
@@ -2,3 +2,4 @@ export * from "./GitBackedRepositorySource.js";
 export * from "./FilesystemRepositorySource.js";
 export * from "./GitHubRepositorySource.js";
 export * from "./GitLabRepositorySource.js";
+export * from "./RefAwareRepositorySource.js";